moved "config" elements into api
